requirements.txt
click
keystoneauth1
python-novaclient
python-glanceclient
python-cinderclient
python-neutronclient
